fix: toggle unit button
